Split the different parts in files
- Also adding a test file for each one.
