+++ src/problems.test.ts
@@ -1,0 +1,19 @@
+import { expect, test } from "bun:test"
+import { JutgeObjectModel as JOM } from "./jom"
+import { describeWithToken } from "./test-utils"
+
+describeWithToken("Problems", () => {
+    const jom = new JOM()
+
+    test.skip("problems.all()", async () => {
+        const problems = await jom.problems.all()
+        console.log("Number of problems:", Object.keys(problems).length)
+        expect(Object.keys(problems).length).toBeGreaterThan(0)
+    })
+
+    test.skip("problems.get()", async () => {
+        const problems = await jom.problems.all()
+        const problemId = Object.keys(problems)[0]
+        const problem = await jom.problems.get(problemId)
+    })
+})
